+++ tests/testthat/test-plot.R
@@ -11,7 +11,7 @@ test_that("ypr_plot_yield", {
 })
 
 test_that("ypr_plot_yield", {
-  gp <- ypr_plot_yield(ypr_population(), y = "Weight",  pi = seq(0, 1, length.out = 10))
+  gp <- ypr_plot_yield(ypr_population(Rk = 10), y = "YPUE",  pi = seq(0, 1, length.out = 10))
   expect_is(gp, "gg")
 })