+++ tests/testthat/test-tabulate.R
@@ -6,10 +6,11 @@ test_that("ypr_tabulate_yield", {
   expect_identical(yield$Type, c("actual", "optimal"))
 
   yields <- ypr_tabulate_yields(ypr_population(), pi = seq(0, 1, length.out = 10))
-  expect_identical(colnames(yields), c("pi", "Yield", "Age", "Length", "Weight"))
+  expect_identical(colnames(yields), c("pi", "Yield", "Age", "Length", "Weight", "Effort"))
   expect_identical(nrow(yields), 10L)
   expect_identical(check_tabulated_yields(yields), yields)
   expect_identical(yields$pi[1:2], c(0,1/9))
+  expect_identical(yields$Effort, yields$pi * 100)
   expect_equal(yields$Yield[1:2], c(0,0.0738), tolerance = 1e-04)
   expect_equal(yields$Weight[1:2], c(NA,3057.662), tolerance = 1e-07)